- attackers follow their target - rocket jammer now has normal attack - fixed double-move when attacking - added floating point range instead of int - kamikaze now destroys angular targets.

+++ open-ig/src/hu/openig/model/BattleGroundTurret.java
@@ -28,7 +28,7 @@ public class BattleGroundTurret {
 	/** The firing sound effect. */
 	public SoundType fire;
 	/** The firing range. */
-	public int maxRange;
+	public double maxRange;
 	/** The damage. */
 	public int damage;
 	/** The rotation time per angle-segment. */
